added docstrings; added block utils unittests
